wip(#1101): extract lower/widget_decl
